Move TelemetryMiddleware to telemetry sub package.
